Added toolbar

+++ app/src/main/java/com/example/msil_cam/MainActivity.kt
@@ -5,13 +5,17 @@ import android.content.Context
 import android.content.Intent
 import android.content.IntentFilter
 import android.os.Bundle
+import android.view.Menu
+import android.view.MenuItem
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.widget.Toolbar
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.datastore.preferences.core.edit
 import androidx.lifecycle.lifecycleScope
 import com.example.msil_cam.databinding.ActivityMainBinding
+import com.example.msil_cam.databinding.ToolbarBinding
 import kotlinx.coroutines.launch
 
 class MainActivity : AppCompatActivity() {
@@ -27,14 +31,28 @@ class MainActivity : AppCompatActivity() {
         mainBinding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(mainBinding.root)
 
-        mainBinding.btnLogout.setOnClickListener {
-            lifecycleScope.launch {
-                dataStore.edit { prefs->
-                    prefs.clear()
+        val toolbar = findViewById<Toolbar>(R.id.toolbar)
+        setSupportActionBar(toolbar)
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
+        menuInflater.inflate(R.menu.toolbar_menu,menu)
+        return true
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        when(item.itemId) {
+            R.id.logout -> {
+                lifecycleScope.launch {
+                    dataStore.edit { prefs->
+                        prefs.clear()
+                    }
+                    startActivity(Intent(this@MainActivity, Login::class.java))
+                    finishAffinity()
                 }
-                startActivity(Intent(this@MainActivity, Login::class.java))
-                finishAffinity()
+                return true
             }
+            else -> return super.onOptionsItemSelected(item)
         }
     }
